remove trace calls
